vulkan_common: Move allocator to the common directory
Allow using the abstraction from the OpenGL backend.
